Update patch set 2
Patch Set 2:

> I haven't dig into the implement, but one suggestion:
 > how about decoupling gating from landing-page? I am thinking of
 > decoupling reporting from landing-page too. IMHP, different purpose
 > should be treated as different project. by doing so to make them
 > modulized, light-weighted, easy-to-maintain, easy-to-migrate and
 > easy-to-reuse. moreover, better not to mix all them together
 > because they are leveraging the same framework.
 > 
 > And as we discussed during ONS, they should be different
 > sub-projects under releng-testresults, to totally decouple them and
 > make our work clear enough to the user, my proposal would be:
 > 
 > releng-testresults/reporting
 > releng-testresults/landingpage
 > releng-testresults/catalog
 > releng-testresults/gating
 > releng-testresults/testapi
 > 
 > I haven't check very much about landingpage and catalog, if you
 > have different opinions, please feel free to discuss.

From my understand, old reporting page, landing page and the new gating reporting page are all OPNFV reporting(offer a friendly view to release manager and other developer), so I put them in the same repos. The old reporting page is a static page, it is hard to implement the filter functions, and the rule among each projects are different, so David said it is hard to understand. For the landing page, actually want to solve the problem in old reporting page, dynamic filter, order by scenario, unified rule. And the new gating reporting page is a new version(will add more function later) of landing page which extends the advantage of landing page and is all designed according David's demands. Once this page is put in use, I will remove the landing page .

Patch-set: 2